Moving debugger logic to its own class and new layout
