+++ host/CMakeLists.txt
@@ -135,7 +135,9 @@ set(PROJECT_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_backend.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_command_queue.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_configuration.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/clem_debugger.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_file_browser.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/clem_frame_state.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_front.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_host_app.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/clem_host_utils.cpp"
